[FEAT] PhotoSearch: 검색바, 정렬 버튼, 목록 UI 구현  #1
