+++ PhotoExplorer/PhotoExplorer/ListConfigurable.swift
@@ -1,0 +1,14 @@
+//
+//  ListConfigurable.swift
+//  PhotoExplorer
+//
+//  Created by 임윤휘 on 1/18/25.
+//
+
+import Foundation
+
+protocol ListConfigurable: AnyObject {
+    associatedtype CellValue
+    
+    func configure(with value: CellValue)
+}
